+++ HISTORY.md
@@ -5,6 +5,7 @@
 **Features and Improvements**
 
 * Added support for DictField - peterfarrell
+* Added support for ListField - peterfarrell
 
 ## 0.6.0 (2020-01-03)